Correct RMS bug cause by a float 'resolution' problem.
Also engine have to be updated, this becuase threshold now need to be given in threshold^2

+++ AxumModuleProcessing.h
@@ -92,7 +92,7 @@ float EQBand5Output[NUMBEROFEQOBJECTS];
 //Dynamics
 float DynamicsRMS[NUMBEROFDYNAMICSOBJECTS];
 volatile float DownwardExpanderRMS[NUMBEROFDYNAMICSOBJECTS];
-volatile float DownwardExpanderAverage[NUMBEROFDYNAMICSOBJECTS];
+volatile double DownwardExpanderAverage[NUMBEROFDYNAMICSOBJECTS];
 float CompressorReductionLevel[NUMBEROFDYNAMICSOBJECTS];
 int DynamicsOn[NUMBEROFDYNAMICSOBJECTS];
